move some part of the training script ine finetune_utils
