+++ scripts/train_classifier.py
@@ -10,69 +10,56 @@ Usage:
 
 Options:
   --path_in=PATH              path to the dataset folder following the structure described in the readme
-  --num_layer_finetune=NUM    Number layer to finetune [default: 2]
+  --num_layer_finetune=NUM    Number layer to finetune, must be integer between 0 and 32 [default: 2]
 
 """
-import torch
-import torch.optim as optim
-import torch.nn as nn
+
 import torch.utils.data
 from docopt import docopt
 import os
-import glob
-import numpy as np
+
 from realtimenet.downstream_tasks.nn_utils import Pipe, LogisticRegression
-from realtimenet import camera
-from realtimenet import engine
+from realtimenet.finetune_utils import training_loops, extract_features, generate_data_loader, evaluation_model
+
 from realtimenet import feature_extractors
 import json
 
 
-class FeaturesDataset(torch.utils.data.Dataset):
-    """Face Landmarks dataset."""
-
-    def __init__(self, files, labels):
-        self.files = files
-        self.labels = labels
-        self.num_timestamp = 5
-
-    def __len__(self):
-        return len(self.files)
-
-    def __getitem__(self, idx):
-        feature = np.load(self.files[idx])
-        num_preds = feature.shape[0]
-        if num_preds <= self.num_timestamp:
-            return self.__getitem__(idx + 1)
-        else:
-            position = np.random.randint(0, num_preds - self.num_timestamp)
-            return [feature[position:position+self.num_timestamp], self.labels[idx]]
-
-def generate_data_loader(features_dir, classes, shuffle=True):
-    features = []
-    labels = []
-    for label in classes:
-        files = os.listdir(os.path.join(features_dir, label))
-        # used to remove .DSstore files on mac
-        feature_temp = [os.path.join(features_dir, label, x) for x in files if not x.startswith('.')]
-        features += feature_temp
-        labels += [class2int[label]] * len(feature_temp)
-    dataset = FeaturesDataset(features, labels)
-    data_loader = torch.utils.data.DataLoader(dataset, shuffle=shuffle, batch_size=64)
-    return data_loader
-
-def uniform_frame_sample(video, sample_rate):
-    """
-    Uniformly sample video frames according to the sample_rate.
-    """
-
-    depth = video.shape[0]
-    if sample_rate < 1.:
-        indices = np.arange(0, depth, 1./sample_rate)
-        offset = int((depth - indices[-1]) / 2)
-        sampled_frames = (indices + offset).astype(np.int32)
-        return video[sampled_frames]
-    return video
+num_layers2timesteps = {
+    0: 1,
+    1: 1,
+    2: 1,
+    3: 1,
+    4: 1,
+    5: 1,
+    6: 1,
+    7: 3,
+    8: 3,
+    9: 5,
+    10: 5,
+    11: 5,
+    12: 7,
+    13: 7,
+    14: 7,
+    15: 9,
+    16: 9,
+    17: 9,
+    18: 19,
+    19: 19,
+    20: 19,
+    21: 21,
+    22: 21,
+    23: 21,
+    24: 21,
+    25: 43,
+    26: 43,
+    27: 43,
+    28: 43,
+    29: 45,
+    30: 45,
+    31: 45,
+    32: 45
+}
 
 if __name__ == "__main__":
     # Parse arguments
@@ -81,6 +68,13 @@ if __name__ == "__main__":
     use_gpu = args['--use_gpu']
     num_layer_finetune = int(args['--num_layer_finetune'])
 
+    # compute the number of timestep necessary for each video features in order to finetune the number of layer wished.
+    num_timestep = num_layers2timesteps.get(int(num_layer_finetune))
+    if not num_timestep:
+        raise NameError('Num layers to finetune not right. Must be integer between 0 and 32.')
+
+    lr_schedule = {0: 0.0001, 10: 0.00001}
+    num_epochs = 20
 
 
     # Load feature extractor
@@ -92,89 +86,28 @@ if __name__ == "__main__":
     # Concatenate feature extractor and met converter
     net = feature_extractor
 
-    # Create inference engine, video streaming and display instances
-    inference_engine = engine.InferenceEngine(net, use_gpu=use_gpu)
-
-
-    postprocessor = []
-    display_ops = []
-
-
-    # list the videos files
+    # list the labels from the training directory
     videos_dir = os.path.join(path_in, "videos_train")
     features_dir = os.path.join(path_in, "features_train")
     classes = os.listdir(videos_dir)
     classes = [x for x in classes if not x.startswith('.')]
 
 
-    # extract features
-    for dataset in ["train", "valid"]:
-        videos_dir = os.path.join(path_in, "videos_" + dataset)
-        features_dir = os.path.join(path_in, "features_" + dataset+ "_" + str(num_layer_finetune))
-        number_videos_found = len(glob.glob(os.path.join(videos_dir, "*", "*.mp4")))
-        number_videos_processed = 0
-        print(f"Found {number_videos_found} videos to process")
-
-        for label in classes:
-            videos = os.listdir(os.path.join(videos_dir, label))
-            videos = [x for x in videos if not x.startswith('.')]
-            for video in videos:
-                number_videos_processed += 1
-                print(f"extract features from video {number_videos_processed} / {number_videos_found}")
-                path_out = os.path.join(features_dir, label, video.replace(".mp4", ".npy"))
-                if os.path.isfile(path_out):
-                    print("features found for this file, skip")
-                else:
-                    file_path = os.path.join(videos_dir, label, video)
-                    path_out = os.path.join(features_dir, label, video.replace(".mp4", ".npy"))
-
-                    video_source = camera.VideoSource(camera_id=None,
-                                                      size=inference_engine.expected_frame_size,
-                                                      filename=file_path,
-                                                      )
-                    frames = []
-                    features = []
-                    video_fps = video_source.get_sample_rate()
-
-                    while True:
-                        images = video_source.get_image()
-                        if images is None:
-                            break
-                        else:
-                            image, image_rescaled = images
-                            frames.append(image_rescaled)
-                    frames = uniform_frame_sample(np.array(frames), 16/video_fps)
-                    clip = np.array([frames]).astype(np.float32)
-
-                    frames = []
-                    try:
-                        if num_layer_finetune > 0:
-                            predictions = inference_engine.process_clip_features_map(clip,
-                                                                                 layer=-num_layer_finetune)
-
-                        else:
-                            predictions = inference_engine.process_clip(clip, training=True)
-                        features = np.array(predictions)
-                        os.makedirs(os.path.dirname((path_out)), exist_ok=True)
-                        np.save(path_out, features)
-                    except:
-                        print("video too short")
-
-
-
-
-
-
     # finetune the model
+    extract_features(path_in, classes, net, num_layer_finetune, use_gpu)
 
     y_train, y_valid = [], []
     X_train, X_valid = [], []
     class2int = {x:e for e,x in enumerate(classes)}
 
+    # create the data loaders
+    trainloader = generate_data_loader(os.path.join(path_in, "features_train_" + str(num_layer_finetune)),
+                                       classes, class2int, num_timesteps=num_timestep)
+    validloader = generate_data_loader(os.path.join(path_in, "features_valid_" + str(num_layer_finetune)),
+                                       classes, class2int, num_timesteps=num_timestep)
 
-    trainloader = generate_data_loader(os.path.join(path_in, "features_train_" + str(num_layer_finetune)), classes)
-    validloader = generate_data_loader(os.path.join(path_in, "features_valid_" + str(num_layer_finetune)), classes)
 
+    # modeify the network to generate the training network on top of the features
     gesture_classifier = LogisticRegression(num_in=feature_extractor.feature_dim,
                                             num_out=len(classes))
     if num_layer_finetune > 0:
@@ -186,105 +119,20 @@ if __name__ == "__main__":
     if use_gpu:
         net = net.cuda()
 
-    criterion = nn.CrossEntropyLoss()
-    optimizer = optim.Adam(net.parameters(), lr=0.0001)
+    training_loops(net, trainloader, validloader, use_gpu, num_epochs=num_epochs, lr_schedule=lr_schedule)
 
-    for epoch in range(80):  # loop over the dataset multiple times
-        if epoch == 40:
-            for param_group in optimizer.param_groups:
-                param_group['lr'] = 0.00001
-
-        running_loss = 0.0
-        net.train()
-        top_pred = []
-        label = []
-        for i, data in enumerate(trainloader, 0):
-            # get the inputs; data is a list of [inputs, labels]
-            inputs, labels = data
-            inputs = inputs.cuda()
-            labels = labels.cuda()
-
-            # zero the parameter gradients
-            optimizer.zero_grad()
-
-            # forward + backward + optimize
-            outputs = []
-            for i in range(len(inputs)):
-                pred = net(inputs[i])
-                outputs.append(pred.mean(dim=0).unsqueeze(0))
-            outputs = torch.cat(outputs, dim=0)
-            loss = criterion(outputs, labels)
-            loss.backward()
-            optimizer.step()
-            label += list(labels.cpu().numpy())
-            top_pred += list(outputs.argmax(dim=1).cpu().numpy())
-
-            # print statistics
-            running_loss += loss.item()
-        percentage = np.sum((np.array(label) == np.array(top_pred))) / len(top_pred)
-        train_top1 = percentage
-        train_loss =  running_loss / len(trainloader)
-
-        running_loss = 0.0
-        top_pred = []
-        label = []
-        net.eval()
-        for i, data in enumerate(validloader, 0):
-            # get the inputs; data is a list of [inputs, labels]
-            with torch.no_grad():
-                inputs, labels = data
-                inputs = inputs.cuda()
-                labels = labels.cuda()
-
-
-                outputs = []
-                for i in range(len(inputs)):
-                    outputs.append(net(inputs[i]).mean(dim=0).unsqueeze(0))
-                outputs = torch.cat(outputs, dim=0)
-
-                label += list(labels.cpu().numpy())
-                top_pred += list(outputs.argmax(dim=1).cpu().numpy())
-
-                # print statistics
-                running_loss += loss.item()
-        percentage = np.sum((np.array(label) == np.array(top_pred))) / len(top_pred)
-        valid_top1 = percentage
-        valid_loss =  running_loss / len(trainloader)
-
-        print('[%d] train loss: %.3f train top1: %.3f valid loss: %.3f top1: %.3f' % (epoch + 1, train_loss, train_top1,
-                                                                                      valid_loss, valid_top1))
-
-    print('Finished Training')
-    print("score on videos")
-    features_dir = os.path.join(path_in, "features_valid_" + str(num_layer_finetune))
-
-    pred = []
-    y = []
-    net.eval()
-    for label in classes:
-        features = os.listdir(os.path.join(features_dir, label))
-        # used to remove .DSstore files on mac
-        features = [x for x in features if not x.startswith('.')]
-        for feature in features:
-            try:
-                feature = np.load(os.path.join(features_dir, label, feature))
-                feature = torch.Tensor(feature).cuda()
-                with torch.no_grad():
-                    output = net(feature).cpu().numpy()
-                top_pred = output.mean(axis=0)
-                pred.append(top_pred.argmax())
-                y.append(class2int[label])
-            except:
-                1
-    pred = np.array(pred)
-    y = np.array(y)
-    percent = (np.sum(pred == y) / len(pred))
-    print(f"top 1 : {percent}")
+    # save the trained model
     if num_layer_finetune > 0:
         state_dict = {**net.feature_extractor.state_dict(), **net.feature_converter.state_dict()}
     else:
         state_dict = net.state_dict()
     torch.save(state_dict, os.path.join(path_in, "classifier.checkpoint"))
-    json.dump(class2int, open(os.path.join(path_in,"class2int.json"), "w"))
+    json.dump(class2int, open(os.path.join(path_in, "class2int.json"), "w"))
+
+    # evaluation score on full videos and not just random temporal crop
+    print("score on videos")
+    features_dir = os.path.join(path_in, "features_valid_" + str(num_layer_finetune))
+    evaluation_model(net, features_dir, classes, class2int, num_timestep, use_gpu)
+
